ReadMqPy/pyproject.toml
[tool.poetry]
name = "readmqpy"
version = "0.1.0"
description = "read Message Queues using Python"
authors = ["ShanKonduru <ShanKonduru@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
kafka-python = "^2.0.2"
Jinja2 = "^3.1.4"
stomp-py = "^8.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
